fix rbcli service command

+++ resources/recipes/prepare_system.rb
@@ -90,13 +90,15 @@ template '/etc/hosts' do
 end
 
 # Build service list for rbcli
-services = node['redborder']['services'] || []
-systemd_services = node['redborder']['systemdservices'] || []
-service_enablement = {}
+unless File.exist?('/etc/redborder/services.json')
+  services = node['redborder']['services'] || []
+  systemd_services = node['redborder']['systemdservices'] || []
+  service_enablement = {}
 
-systemd_services.each do |service_name, systemd_name|
-  service_enablement[systemd_name.first] = services[service_name]
+  systemd_services.each do |service_name, systemd_name|
+    service_enablement[systemd_name.first] = services[service_name]
+  end
+
+  Chef::Log.info('Saving services enablement into /etc/redborder/services.json')
+  File.write('/etc/redborder/services.json', JSON.pretty_generate(service_enablement))
 end
-
-Chef::Log.info('Saving services enablement into /etc/redborder/services.json')
-File.write('/etc/redborder/services.json', JSON.pretty_generate(service_enablement))
